fixed issue when deleting watch with no image

+++ chronos/watch/views.py
@@ -154,8 +154,9 @@ def delete_watch(request, pk):
     if request.method == 'POST':
         form = DeleteWatchForm(request.POST, request.FILES, instance=watch)
         if form.is_valid():
-            image_path = watch.image.path
-            os.remove(image_path)
+            if watch.image:
+                image_path = watch.image.path
+                os.remove(image_path)
             watch.delete()
             messages.success(request, WATCH_DELETE_SUCCESS_MESSAGE)
             return redirect('show dashboard')
